+++ src/revolut/models/__init__.py
@@ -9,15 +9,22 @@ from .common import (
     RevolutModel,
 )
 from .customer import Customer
+from .dispute import Dispute
+from .location import Location
 from .order import Order, OrderList
 from .payment import Payment
 from .payment_method import PaymentMethod
+from .payout import Payout
 from .refund import Refund
+from .report_run import ReportRun
+from .subscription import Subscription
 from .webhook import Webhook, WebhookEvent, WebhookEventPayload
 
 __all__ = [
     "CaptureMode",
     "Customer",
+    "Dispute",
+    "Location",
     "Money",
     "Order",
     "OrderList",
@@ -26,8 +33,11 @@ __all__ = [
     "Payment",
     "PaymentMethod",
     "PaymentState",
+    "Payout",
     "Refund",
+    "ReportRun",
     "RevolutModel",
+    "Subscription",
     "Webhook",
     "WebhookEvent",
     "WebhookEventPayload",