+++ src/revolut/resources/__init__.py
@@ -1,23 +1,38 @@
 """Resource classes for each Merchant API object group."""
 
 from .customers import AsyncCustomersResource, CustomersResource
+from .disputes import AsyncDisputesResource, DisputesResource
+from .locations import AsyncLocationsResource, LocationsResource
 from .orders import AsyncOrdersResource, OrdersResource
 from .payment_methods import AsyncPaymentMethodsResource, PaymentMethodsResource
 from .payments import AsyncPaymentsResource, PaymentsResource
+from .payouts import AsyncPayoutsResource, PayoutsResource
 from .refunds import AsyncRefundsResource, RefundsResource
+from .report_runs import AsyncReportRunsResource, ReportRunsResource
+from .subscriptions import AsyncSubscriptionsResource, SubscriptionsResource
 from .webhooks import AsyncWebhooksResource, WebhooksResource
 
 __all__ = [
     "AsyncCustomersResource",
+    "AsyncDisputesResource",
+    "AsyncLocationsResource",
     "AsyncOrdersResource",
     "AsyncPaymentMethodsResource",
     "AsyncPaymentsResource",
+    "AsyncPayoutsResource",
     "AsyncRefundsResource",
+    "AsyncReportRunsResource",
+    "AsyncSubscriptionsResource",
     "AsyncWebhooksResource",
     "CustomersResource",
+    "DisputesResource",
+    "LocationsResource",
     "OrdersResource",
     "PaymentMethodsResource",
     "PaymentsResource",
+    "PayoutsResource",
     "RefundsResource",
+    "ReportRunsResource",
+    "SubscriptionsResource",
     "WebhooksResource",
 ]